Setting userDetail page (showing own video)

+++ controllers/videoController.js
@@ -44,7 +44,7 @@ export const postUpload = async (req, res) => {
     });
     req.user.videos.push(newVideo.id);
     req.user.save();
-    console.log(newVideo);
+    
    // To Do Upload and Save video
    res.redirect(routes.videoDetail(newVideo.id));
 }
@@ -57,7 +57,7 @@ export const videoDetail =  async (req, res) => {
         const video = await Video.findById(id).populate("creator");
         //inside of populate type is only object id
 
-        console.log(video);
+        
         res.render("videoDetail",{pageTitle:video.title, video});
     }catch(error){
         console.log(error);
@@ -102,9 +102,14 @@ export const deleteVideo = async (req, res) => {
         params : {id}
     }=req;
     try{
-        await Video.findOneAndRemove({_id : id});
+        const video = await Video.findById(id);
+        if(video.creator !== req.user.id){
+            throw Error();
+        } else {
+            await Video.findOneAndRemove({_id : id});
+        }
     } catch(error) {
         console.log(error);
     }
     res.redirect(routes.home);
-}
+};
